+++ dhall.nix
@@ -1,20 +1,19 @@
 { mkDerivation, aeson, aeson-pretty, ansi-terminal, base
 , bytestring, case-insensitive, cborg, cborg-json, containers
 , contravariant, criterion, cryptonite, deepseq, Diff, directory
-, doctest, dotgen, exceptions, filepath, haskeline, http-client
-, http-client-tls, http-types, lens-family-core, megaparsec, memory
-, mockery, mtl, optparse-applicative, parsers, prettyprinter
-, prettyprinter-ansi-terminal, QuickCheck, quickcheck-instances
-, repline, scientific, serialise, stdenv, tasty, tasty-hunit
-, tasty-quickcheck, template-haskell, text, transformers
-, unordered-containers, uri-encode, vector
+, doctest, dotgen, exceptions, filepath, foldl, haskeline
+, http-client, http-client-tls, http-types, lens-family-core
+, megaparsec, memory, mockery, mtl, optparse-applicative, parsers
+, prettyprinter, prettyprinter-ansi-terminal, QuickCheck
+, quickcheck-instances, repline, scientific, serialise, stdenv
+, tasty, tasty-hunit, tasty-quickcheck, template-haskell, text
+, transformers, transformers-compat, turtle, unordered-containers
+, uri-encode, vector
 }:
 mkDerivation {
   pname = "dhall";
-  version = "1.20.1";
-  sha256 = "e077e8f4945484db4e35e8ae422e9eabac1b155972602f0404d818e3e185bcdc";
-  revision = "1";
-  editedCabalFile = "1km0zbbahhq24s84s9gcck1javhplqjg51q4qf8i19iahnxkl3rq";
+  version = "1.23.0";
+  sha256 = "eda7b9d1baad8214f83aaf7e7ce5e374c32a62f58ca69734024fb3f254bc9d1c";
   isLibrary = true;
   isExecutable = true;
   libraryHaskellDepends = [
@@ -24,14 +23,15 @@ mkDerivation {
     http-types lens-family-core megaparsec memory mtl
     optparse-applicative parsers prettyprinter
     prettyprinter-ansi-terminal repline scientific serialise
-    template-haskell text transformers unordered-containers uri-encode
-    vector
+    template-haskell text transformers transformers-compat
+    unordered-containers uri-encode vector
   ];
   executableHaskellDepends = [ base ];
   testHaskellDepends = [
-    base containers deepseq directory doctest filepath mockery
-    prettyprinter QuickCheck quickcheck-instances serialise tasty
-    tasty-hunit tasty-quickcheck text transformers vector
+    base bytestring cborg containers deepseq directory doctest filepath
+    foldl mockery prettyprinter QuickCheck quickcheck-instances
+    serialise tasty tasty-hunit tasty-quickcheck text transformers
+    turtle vector
   ];
   benchmarkHaskellDepends = [
     base bytestring containers criterion directory serialise text